fix: 调整 next-sdk 的打包产物，支持内联css,svg
